Backend/requirements.txt
asgiref==3.3.1
black==21.11b1
boto3==1.21.30
botocore==1.24.46
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-admin-honeypot==1.1.0
django-autoslug==1.9.8
django-cors-headers==3.6.0
django-filter==21.1
django-ipware==3.0.2
django-mssql-backend==2.8.1
django-otp==1.0.6
django-storages==1.12.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-recaptcha==2.0.4
drf-writable-nested==0.6.3
drf-yasg==1.20.0
facebook-sdk==3.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==1.0.1
MarkupSafe==1.1.1
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
mypy-extensions==1.0.0
oauthlib==3.1.0
packaging==20.8
pathspec==0.12.1
Pillow==9.2.0
platformdirs==4.2.2
protobuf==3.17.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyodbc==5.1.0
pyparsing==2.4.7
pytest==2.9.1
python-dateutil==2.9.0.post0
python-decouple==3.4
python-magic-bin==0.4.14
python-twitter==3.5
pytz==2020.5
qrcode==6.1
regex==2024.7.24
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.5.2
six==1.15.0
sqlparse==0.4.1
tomli==1.2.3
typing-extensions==4.12.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
